Rename 'utils' to '_utils'
And update import style to be consistently absolute FromImport.
