+++ src/globus_sdk/services/auth/flow_managers/authorization_code.py
@@ -4,8 +4,8 @@ import logging
 import typing as t
 import urllib.parse
 
-from globus_sdk import utils
 from globus_sdk._types import ScopeCollectionType
+from globus_sdk._utils import slash_join
 
 from .._common import stringify_requested_scopes
 from ..response import OAuthAuthorizationCodeResponse
@@ -86,7 +86,7 @@ class GlobusAuthorizationCodeFlowManager(GlobusOAuthFlowManager):
         either to your provided ``redirect_uri`` or to the default location,
         with the ``auth_code`` embedded in a query parameter.
         """
-        authorize_base_url = utils.slash_join(
+        authorize_base_url = slash_join(
             self.auth_client.base_url, "/v2/oauth2/authorize"
         )
         log.debug(f"Building authorization URI. Base URL: {authorize_base_url}")
