+++ src/globus_sdk/services/auth/flow_managers/native_app.py
@@ -8,8 +8,8 @@ import re
 import typing as t
 import urllib.parse
 
-from globus_sdk import utils
 from globus_sdk._types import ScopeCollectionType
+from globus_sdk._utils import slash_join
 from globus_sdk.exc import GlobusSDKUsageError
 
 from .._common import stringify_requested_scopes
@@ -127,7 +127,7 @@ class GlobusNativeAppFlowManager(GlobusOAuthFlowManager):
         # default to `/v2/web/auth-code` on whatever environment we're looking
         # at -- most typically it will be `https://auth.globus.org/`
         self.redirect_uri = redirect_uri or (
-            utils.slash_join(auth_client.base_url, "/v2/web/auth-code")
+            slash_join(auth_client.base_url, "/v2/web/auth-code")
         )
 
         # make a challenge and secret to keep
@@ -166,7 +166,7 @@ class GlobusNativeAppFlowManager(GlobusOAuthFlowManager):
         either to your provided ``redirect_uri`` or to the default location,
         with the ``auth_code`` embedded in a query parameter.
         """
-        authorize_base_url = utils.slash_join(
+        authorize_base_url = slash_join(
             self.auth_client.base_url, "/v2/oauth2/authorize"
         )
         log.debug(f"Building authorization URI. Base URL: {authorize_base_url}")
