+++ src/globus_sdk/services/timers/data.py
@@ -8,10 +8,10 @@ import typing as t
 
 from globus_sdk._missing import MISSING, MissingType
 from globus_sdk._payload import Payload
+from globus_sdk._utils import slash_join
 from globus_sdk.config import get_service_url
 from globus_sdk.exc import warn_deprecated
 from globus_sdk.services.transfer import TransferData
-from globus_sdk.utils import slash_join
 
 log = logging.getLogger(__name__)
