+++ tests/functional/services/timers/test_jobs.py
@@ -3,8 +3,9 @@ import json
 
 import pytest
 
-from globus_sdk import TimerJob, TimersAPIError, TransferData, config, exc, utils
+from globus_sdk import TimerJob, TimersAPIError, TransferData, config, exc
 from globus_sdk._testing import get_last_request, load_response
+from globus_sdk._utils import slash_join
 from tests.common import GO_EP1_ID, GO_EP2_ID
 
 
@@ -60,7 +61,7 @@ def test_create_job(client, start, interval):
         assert req_body["interval"] == interval.total_seconds()
     else:
         assert req_body["interval"] == interval
-    assert req_body["callback_url"] == utils.slash_join(
+    assert req_body["callback_url"] == slash_join(
         config.get_service_url("actions"), "/transfer/transfer/run"
     )
